pyproject.toml
[tool.poetry]
name = "Dungeon Generator"
version = "0.1.0"
description = "Generating dungeons using FantomX11 syntax"
authors = ["Pedro-Juan Ferrer <vehrka@gmail.com>"]
license = "Beerware"

[tool.poetry.dependencies]
python = "^3.8"
python-igraph = "^0.9.8"

[tool.poetry.dev-dependencies]
ipython = "^7.29.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
